Kode der kan hente co2 data fra url
