+++ assets/scss/calculateSection/_CO2Result.scss
@@ -1,0 +1,15 @@
+@use "../variables";
+@use "../mixins";
+@use "../font";
+@use "../placeholder";
+
+#calculate-result {
+  position: absolute;
+  top: 100%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  z-index: 999;
+  @include mixins.calculateFlexbox(column, center, center, 0.7rem);
+  width: 70%;
+  background: variables.$background_dark;
+}
