+++ assets/scss/style.scss
@@ -5,3 +5,4 @@
 @use "./section";
 @use "./footer";
 @use "../scss/calculateSection/calculate.scss";
+@use "../scss/calculateSection/CO2Result.scss";
